+++ tests/conftest.py
@@ -9,7 +9,6 @@ from urllib.parse import urljoin
 import asyncpg
 import pytest
 from fastapi import APIRouter
-from fastapi.responses import ORJSONResponse
 from httpx import ASGITransport, AsyncClient
 from pypgstac import __version__ as pgstac_version
 from pypgstac.db import PgstacDB
@@ -18,6 +17,7 @@ from pytest_postgresql.janitor import DatabaseJanitor
 from stac_fastapi.api.app import StacApi
 from stac_fastapi.api.models import (
     ItemCollectionUri,
+    JSONResponse,
     create_get_request_model,
     create_post_request_model,
     create_request_model,
@@ -189,7 +189,7 @@ def api_client(request):
         search_get_request_model=search_get_request_model,
         search_post_request_model=search_post_request_model,
         collections_get_request_model=collection_search_extension.GET,
-        response_class=ORJSONResponse,
+        response_class=JSONResponse,
         router=APIRouter(prefix=prefix),
         health_check=health_check,
     )
@@ -290,14 +290,11 @@ async def load_test2_item(app_client, load_test_data, load_test2_collection):
     return Item.model_validate(resp.json())
 
 
-@pytest.fixture(
-    scope="session",
-)
-def api_client_no_ext():
-    api_settings = Settings(
-        testing=True,
-    )
-    return StacApi(
+@pytest.fixture(scope="function")
+async def app_no_ext(database):
+    """Default stac-fastapi-pgstac application without only the transaction extensions."""
+    api_settings = Settings(testing=True)
+    api_client_no_ext = StacApi(
         settings=api_settings,
         extensions=[
             TransactionExtension(client=TransactionsClient(), settings=api_settings)
@@ -306,9 +303,6 @@ def api_client_no_ext():
         health_check=health_check,
     )
 
-
-@pytest.fixture(scope="function")
-async def app_no_ext(api_client_no_ext, database):
     postgres_settings = PostgresSettings(
         postgres_user=database.user,
         postgres_pass=database.password,
@@ -319,12 +313,9 @@ async def app_no_ext(api_client_no_ext, database):
     )
     logger.info("Creating app Fixture")
     time.time()
-    app = api_client_no_ext.app
-    await connect_to_db(app, postgres_settings=postgres_settings)
-
-    yield app
-
-    await close_db_connection(app)
+    await connect_to_db(api_client_no_ext.app, postgres_settings=postgres_settings)
+    yield api_client_no_ext.app
+    await close_db_connection(api_client_no_ext.app)
 
     logger.info("Closed Pools.")
 
@@ -334,5 +325,72 @@ async def app_client_no_ext(app_no_ext):
     logger.info("creating app_client")
     async with AsyncClient(
         transport=ASGITransport(app=app_no_ext), base_url="http://test"
+    ) as c:
+        yield c
+
+
+@pytest.fixture(scope="function")
+async def app_no_transaction(database):
+    """Default stac-fastapi-pgstac application without any extensions."""
+    api_settings = Settings(testing=True)
+    api = StacApi(
+        settings=api_settings,
+        extensions=[],
+        client=CoreCrudClient(),
+        health_check=health_check,
+    )
+
+    postgres_settings = PostgresSettings(
+        postgres_user=database.user,
+        postgres_pass=database.password,
+        postgres_host_reader=database.host,
+        postgres_host_writer=database.host,
+        postgres_port=database.port,
+        postgres_dbname=database.dbname,
+    )
+    logger.info("Creating app Fixture")
+    time.time()
+    await connect_to_db(api.app, postgres_settings=postgres_settings)
+    yield api.app
+    await close_db_connection(api.app)
+
+    logger.info("Closed Pools.")
+
+
+@pytest.fixture(scope="function")
+async def app_client_no_transaction(app_no_transaction):
+    logger.info("creating app_client")
+    async with AsyncClient(
+        transport=ASGITransport(app=app_no_transaction), base_url="http://test"
+    ) as c:
+        yield c
+
+
+@pytest.fixture(scope="function")
+async def default_app(database, monkeypatch):
+    """Test default stac-fastapi-pgstac application."""
+    monkeypatch.setenv("POSTGRES_USER", database.user)
+    monkeypatch.setenv("POSTGRES_PASS", database.password)
+    monkeypatch.setenv("POSTGRES_HOST_READER", database.host)
+    monkeypatch.setenv("POSTGRES_HOST_WRITER", database.host)
+    monkeypatch.setenv("POSTGRES_PORT", str(database.port))
+    monkeypatch.setenv("POSTGRES_DBNAME", database.dbname)
+    monkeypatch.delenv("ENABLED_EXTENSIONS", raising=False)
+
+    monkeypatch.setenv("ENABLE_TRANSACTIONS_EXTENSIONS", "TRUE")
+    monkeypatch.setenv("USE_API_HYDRATE", "TRUE")
+    monkeypatch.setenv("ENABLE_RESPONSE_MODELS", "TRUE")
+
+    from stac_fastapi.pgstac.app import app
+
+    await connect_to_db(app)
+    yield app
+    await close_db_connection(app)
+
+
+@pytest.fixture(scope="function")
+async def default_client(default_app):
+    async with AsyncClient(
+        transport=ASGITransport(app=default_app), base_url="http://test"
     ) as c:
         yield c